# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hospital Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /entities:
    get:
      operationId: listEntities
      tags: [entities]
      summary: List all legal entities
      responses:
        "200":
          description: Entity list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Entity"
    post:
      operationId: createEntity
      tags: [entities]
      summary: Create new entity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEntityBody"
      responses:
        "201":
          description: Created entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entity"
  /entities/{id}:
    put:
      operationId: updateEntity
      tags: [entities]
      summary: Update entity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEntityBody"
      responses:
        "200":
          description: Updated entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entity"
...
  /doctors:
    get:
      operationId: listDoctors
      tags: [doctors]
      summary: List all doctors
      responses:
        "200":
          description: Doctor list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Doctor"
    post:
      operationId: createDoctor
      tags: [doctors]
      summary: Add a doctor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDoctorBody"
      responses:
        "201":
          description: Created doctor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Doctor"
  /doctors/{id}:
    get:
      operationId: getDoctor
      tags: [doctors]
      summary: Get doctor by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Doctor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Doctor"
    put:
      operationId: updateDoctor
      tags: [doctors]
      summary: Update doctor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDoctorBody"
      responses:
        "200":
          description: Updated doctor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Doctor"

  # ─── PRESCRIPTION TEMPLATES ─────────────────────────────────
  /prescription-templates:
    get:
      operationId: listPrescriptionTemplates
      tags: [prescription-templates]
      summary: List prescription templates
      parameters:
        - name: doctorId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Templates list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PrescriptionTemplate"
    post:
      operationId: createPrescriptionTemplate
      tags: [prescription-templates]
      summary: Create a prescription template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePrescriptionTemplateBody"
      responses:
        "201":
          description: Created template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrescriptionTemplate"
  /prescription-templates/{id}:
    get:
      operationId: getPrescriptionTemplate
      tags: [prescription-templates]
      summary: Get template by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrescriptionTemplate"
    put:
      operationId: updatePrescriptionTemplate
      tags: [prescription-templates]
      summary: Update template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePrescriptionTemplateBody"
      responses:
        "200":
          description: Updated template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrescriptionTemplate"
    delete:
      operationId: deletePrescriptionTemplate
      tags: [prescription-templates]
      summary: Delete template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── OPD ────────────────────────────────────────────────────
...
components:
  schemas:
...
    Entity:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        owner:
          type: string
        gstin:
          type: string
          nullable: true
        pan:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        mobile:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, name, type, owner]
    CreateEntityBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        owner:
          type: string
        gstin:
          type: string
        pan:
          type: string
        address:
          type: string
        mobile:
          type: string
        email:
          type: string
      required: [name, type, owner]
...
    Doctor:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        specialization:
          type: string
        qualification:
          type: string
        phone:
          type: string
        email:
          type: string
        registrationNo:
          type: string
        signatureUrl:
          type: string
        opdFee:
          type: number
        isActive:
          type: boolean
      required: [id, name, specialization]
    CreateDoctorBody:
      type: object
      properties:
        name:
          type: string
        specialization:
          type: string
        qualification:
          type: string
        phone:
          type: string
        email:
          type: string
        registrationNo:
          type: string
        signatureUrl:
          type: string
        opdFee:
          type: number
        isActive:
          type: boolean
      required: [name, specialization]
    PrescriptionTemplate:
      type: object
      properties:
        id:
          type: integer
        doctorId:
          type: integer
        name:
          type: string
        chiefComplaints:
          type: string
        diagnosis:
          type: string
        medicines:
          type: array
          items:
            $ref: "#/components/schemas/TemplateMedicine"
        labTests:
          type: string
        radiologyTests:
          type: string
        advise:
          type: string
        specialAdvise:
          type: string
        followUpDays:
          type: integer
        createdAt:
          type: string
      required: [id, name]
    TemplateMedicine:
      type: object
      properties:
        medicineId:
          type: integer
        medicineName:
          type: string
        genericName:
          type: string
        dose:
          type: string
        timing:
          type: string
        frequency:
          type: string
        duration:
          type: string
        qty:
          type: number
        route:
          type: string
        notes:
          type: string
        instructions:
          type: string
      required: [medicineName]
    CreatePrescriptionTemplateBody:
      type: object
      properties:
        doctorId:
          type: integer
        name:
          type: string
        chiefComplaints:
          type: string
        diagnosis:
          type: string
        medicines:
          type: array
          items:
            $ref: "#/components/schemas/TemplateMedicine"
        labTests:
          type: string
        radiologyTests:
          type: string
        advise:
          type: string
        specialAdvise:
          type: string
        followUpDays:
          type: integer
      required: [name]
...